+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Fix missing go path for checkmake
+
 ## 3.8.0 - 2025-01-24
 ### Added
 - Add checkmake installation